Moving models from contract/ -> subgraph/
